Continue module decomposition: extract auth templates from templates.js
Move auth-focused template functions into templates/auth-templates.js and keep templates.js as a stable facade exporting the same public API.

Preserve rendered auth page output and existing route/template contracts while reducing the main template hotspot.
